fix : softmax, cross_entropy 수정
softmax 입력 1차원 -> 2차원으로 수정
forward -> call로 변경
